Refactor App, move controllers to dedicated files
