+++ src/app/controllers/mod.rs
@@ -1,0 +1,7 @@
+pub mod app_controller;
+pub mod mod_list_controller;
+pub mod install_controller;
+
+pub use app_controller::AppController;
+pub use mod_list_controller::ModListController;
+pub use install_controller::InstallController;
